requirements.txt
python-telegram-bot==21.1.1
requests==2.28.1
validators==0.20.0